requirements.txt
flask==2.0.1
opencv-python==4.8.0
numpy==1.24.0
Pillow==10.0.0
google-generativeai==0.3.0
python-dotenv==1.0.0
Werkzeug==2.0.1